renderer: Switch order of demos.
More sensible structure.
