+++ assets/test/shaders/demo_0_display.frag.glsl
@@ -1,10 +1,7 @@
 #version 330
 
-uniform sampler2D color_texture;
-
-in vec2 v_uv;
 out vec4 col;
 
 void main() {
-	col = texture(color_texture, v_uv);
+    col = vec4(1.0, 0.4, 0.0, 0.8);
 }
