+++ assets/test/shaders/demo_1_display.vert.glsl
@@ -5,6 +5,6 @@ layout(location=1) in vec2 uv;
 out vec2 v_uv;
 
 void main() {
-	gl_Position = vec4(position, 0.0, 1.0);
+	gl_Position =  vec4(position, 0.0, 1.0);
 	v_uv = uv;
 }
