+++ assets/test/shaders/demo_1_obj.frag.glsl
@@ -9,23 +9,9 @@ layout(location=1) out uint id;
 
 void main() {
 	vec4 tex_val = texture(tex, v_uv);
-	int alpha = int(round(tex_val.a * 255));
-	switch (alpha) {
-		case 0:
+	if (tex_val.a == 0) {
 		discard;
-		break;
-		case 254:
-		col = vec4(1.0f, 0.0f, 0.0f, 1.0f);
-		break;
-		case 252:
-		col = vec4(0.0f, 1.0f, 0.0f, 1.0f);
-		break;
-		case 250:
-		col = vec4(0.0f, 0.0f, 1.0f, 1.0f);
-		break;
-		default:
-	    col = tex_val;
-		break;
 	}
+	col = tex_val;
 	id = u_id;
 }
