+++ assets/test/shaders/demo_2_display.frag.glsl
@@ -1,7 +1,10 @@
 #version 330
 
+uniform sampler2D color_texture;
+
+in vec2 v_uv;
 out vec4 col;
 
 void main() {
-    col = vec4(1.0, 0.4, 0.0, 0.8);
+	col = texture(color_texture, v_uv);
 }
