+++ libopenage/renderer/demo/demo_2.h
@@ -7,10 +7,14 @@
 namespace openage::renderer::tests {
 
 /**
- * Basic test for creating a window and displaying a single-color texture
- * created by a shader.
+ * Show loading of .sprite and .texture meta information files and display
+ * of subtextures from a texture atlas:
+ *     - Window creation
+ *     - Loading shaders
+ *     - Loading resource files (.sprite and .texture)
+ *     - Rendering of subtextures from a texture atlas (analogue to demo 1)
  *
- * @param path Path to the project rootdir.
+ * @param path Path to the openage asset directory.
  */
 void renderer_demo_2(const util::Path &path);
